requirements.txt
django==4.2.1
djangorestframework==3.14.0
numpy==1.23.5
pillow==9.5.0
tensorflow==2.12.0
tensorflow_hub==0.13.0
django-cors-headers==4.0.0
opencv-python==4.7.0.72